5.python-pokemon-battle/requirements.txt
Flask==3.1.0
Werkzeug==3.1.3
Jinja2==3.1.3
itsdangerous==2.2.0
pymongo==4.0.1
python-dotenv==1.1.0
Flask-Cors==5.0.1
click==8.1.8
